+++ src/main.ts
@@ -18,7 +18,7 @@ import i18next from "i18next";
 import {getTitleField, regexOnFileName} from "./conversion/file_path";
 import { resources, translationLanguage } from "./i18n/i18next";
 import {migrateSettings} from "./settings/migrate";
-import {ChooseWhichRepoToRun} from "./settings/modals/commands_modals";
+import {ChooseWhichRepoToRun} from "./commands/suggest_other_repo_commands_modal";
 import {
 	createLinkCommand,
 	purgeNotesRemoteCommand,
